Improved memory usage

+++ src/microControl/Cerebrum.h
@@ -40,7 +40,7 @@ private:
 
     double movementSpeed = .75, preciseMovementSpeed = -.25, rampMovementSpeed = .5;
  	uint8_t distanceMoveTiles = 30, turnCounter = 0, mazeFloor = 0, wallTolerance = 9, wallThreshold = 15;
-	int  angles[4];
+	short int  angles[4];
  	bool hasStarted = false;
 };
 #endif
