+++ src/microControl/Node.cpp
@@ -6,21 +6,21 @@ Node::Node(){
     this->g = 0;
 }
 
-Node::Node(int x, int y) : Coord(x, y){
+Node::Node(uint8_t x, uint8_t y) : Coord(x, y){
     this->h = 0;
     this->f = 0;    
     this->g = 0;
 }
 
-void Node::setF(int f){
+void Node::setF(uint8_t f){
     this->f = f;
 }
 
-void Node::setH(int h){
+void Node::setH(uint8_t h){
     this->h = h;
 }
 
-void Node::setG(int g){
+void Node::setG(uint8_t g){
     this->g = g;
 }
 
@@ -28,15 +28,15 @@ void Node::setPrevious(Node* previous){
     this->previous = previous;
 }
 
-int Node::getF() const{
+uint8_t Node::getF() const{
     return f;
 }
 
-int Node::getH() const{
+uint8_t Node::getH() const{
     return h;
 }
 
-int Node::getG() const{
+uint8_t Node::getG() const{
     return g;
 }
 
@@ -50,14 +50,14 @@ void Node::addNeighbor(Node* neighbor){
     }
 }
 
-Node* Node::getNeighbors(int index){
+Node* Node::getNeighbors(uint8_t index){
     if(index < 4)
         return neighbors[index];
     else 
         return nullptr;
 }
 
-int Node::getCurrentNeighbors(){
+uint8_t Node::getCurrentNeighbors(){
     return currentNeighbors;
 }
