+++ src/microControl/Path.cpp
@@ -8,13 +8,13 @@ Path::Path(vector<Coord> path){
     this->path = path;
 }
 
-Coord Path::getCoordAt(int index){
+Coord Path::getCoordAt(uint8_t index){
     if(index >= 0 && index < path.size()){
         return path[index];
     }
 }
 
-int Path::getLength(){
+uint8_t Path::getLength(){
     return path.size();
 }
