+++ src/microControl/Path.h
@@ -1,15 +1,15 @@
 #ifndef __PATH_H_
 #define __PATH_H_
 #include "Coord.h"
-#include <vector>
+#include <ArduinoSTL.h>
 using namespace std;
 
 class Path {
 public:
 	Path();
 	Path(vector<Coord> path);
-	Coord getCoordAt(int index);
-	int getLength();
+	Coord getCoordAt(uint8_t index);
+	uint8_t getLength();
 	vector<Coord> getPath();
 private:
 	vector<Coord> path;
